Some changes to get the sound synthesis part working.
Also added a python script to the plate directory to automate modal
sound precomputation (at least partially)

+++ src/linearalgebra/SPARSE_MATRIX.h
@@ -243,7 +243,7 @@ class SPARSE_MATRIX {
             {
                 size_t totalNZ = 0;
 
-                for ( int col_idx = 0; col_idx < _ncol; col_idx++ ) {
+                for ( int col_idx = 0; col_idx < (int)_ncol; col_idx++ ) {
                     for ( int row_ptr = _p[ col_idx ]; row_ptr < _p[ col_idx + 1 ];
                             row_ptr++ )
                     {
